AzureUtilities/ReadBlobData/requirements.txt
azure-core==1.24.0
azure-identity==1.10.0
azure-storage-blob==12.12.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
idna==3.3
isodate==0.6.1
msal==1.17.0
msal-extensions==1.0.0
msrest==0.6.21
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.2
portalocker==2.4.0
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pywin32==304
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
wincertstore==0.2
